Clean up Papyrus and fix invalid handle
There's no need to prefix types with `Papyrus` in the `papyrus`
namespace. In addition, `usingnamespace` has been removed in cases where
it's used to pollute the namespace, prefering to explicitly shorten the
path to a declaration (see `NodeHandle`). Now it will be significantly
easier to navigate the module from the perspective of an outsider.

There was also an issue that I ran into with deleting nodes. It seems
that `MousePick` would remember the last interacted node, however if
this node was removed it would still attempt to use it in some
scenarios. I added a check to make sure the node was valid but I'm not
sure if that's the best solution here.